Initial upgrade of most code files
Mostly done simulation parameters

Looked through the auto-evo file changes

Updated the tutorial folder files

Handled tools folder files

Initial syntax conversion of thriveopedia, has broken method calls

Handled the steam folder files

Handle the stage starters

Initial syntax conversion of space stage files

Syntax conversion of society stage files

Mostly just basic syntax upgrade on the saving folder

Fixed native interop C# side files

Syntax upgrade on the modding folder

Upgraded benchmark files

Upgraded late multicellular folder

Initial upgrade of industrial stage

Upgrade of early multicellular files

Basic converion of awakening stage

Basic conversion of ascension files

Finished first pass of fixing up engine folder  files

First full round of converting the general folder files

First part of gui_common file upgrades

split as there's some really tricky files here to convert

Auto updated font files

Rest of GUI common initial conversion

Fixe name / deletions

Autogenerated font import settings

Part of the microbe stage files upgraded

One more group of microbe upgrade files done as a first draft

probably one more commit to fully handle all the microbe files

And the microbe systems

this is really the last thing for today

Here's the last microbe stage stuff initially converted

didn't turn out that I could resist and stop after the previous commit

Few more small leftover changes I had

Updated a bunch of various stuff to compile

some with bulk renames and partially manually

+++ src/engine/input/RunOnKeyChangeAttribute.cs
@@ -33,7 +33,7 @@ public class RunOnKeyChangeAttribute : RunOnKeyAttribute
         return CallMethod(cachedMethodCallParameters!);
     }
 
-    public override void OnProcess(float delta)
+    public override void OnProcess(double delta)
     {
     }
 }
